training/requirements.txt
kaggle
scipy
pillow
absl-py
numpy
